Fix xmlwf for XML_UNICODE and !XML_UNICODE_WCHAR_T
This is a large, horrible commit.  Unfortunately I can't see a way
of making it less horrible, and nothing smaller even comes close to
compiling.

We add (limited) support for 16-bit character strings and for
separating the environment (which can be wchar_t under Windows if
you want) and the library's internal representation (XML_Char).

Note the underlying library is not guaranteed correct at this point.